+++ Project.toml
@@ -8,15 +8,16 @@ CUDAdrv = "c5f51814-7f29-56b8-a69c-e4d8f6be1fde"
 CUDAnative = "be33ccc6-a3ff-5ff2-a52e-74243cff1e17"
 CuArrays = "3a865a2d-5b23-5a0f-bc46-62713ec82fae"
 DiffEqBase = "2b5f629d-d688-5b77-993f-72d75c75574e"
+Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
 GPUifyLoops = "ba82f77b-6841-5d2e-bd9f-4daf811aec27"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 
 [compat]
+CUDAdrv = "4.0"
+CUDAnative = "2.5"
 CuArrays = "1.4"
 DiffEqBase = "6.5"
 GPUifyLoops = "0.2"
-CUDAdrv = "4.0"
-CUDAnative = "2.5"
 OrdinaryDiffEq = "5.23"
 julia = "1"
